feat: Refactor Accordion component with new hooks and styles for improved functionality and performance
