+++ packages/native/src/components/accordion/accordion-item.tsx
@@ -1,19 +1,13 @@
-import React, { useRef, useEffect, useState, useMemo, useCallback } from 'react'
+import React from 'react'
+import { View, Text, Pressable, Animated } from 'react-native'
+import { styles } from './accordion-item.style'
 import {
-  View,
-  Text,
-  Pressable,
-  StyleSheet,
-  Animated,
-  Easing,
-  type LayoutChangeEvent,
-  type TextStyle,
-  type ViewStyle,
-} from 'react-native'
-import { useXUITheme } from '../../core'
-import { useAccordionContext } from './accordion-context'
-import type { AccordionItemProps } from './accordion-item.type'
+  useAccordionItemState,
+  useAccordionItemAnimation,
+  useAccordionItemStyles,
+} from './accordion-item.hook'
 import { ChevronRightIcon } from './chevron-right-icon'
+import type { AccordionItemProps } from './accordion-item.type'
 
 export const AccordionItem: React.FC<AccordionItemProps> = ({
   itemKey,
@@ -32,146 +26,32 @@ export const AccordionItem: React.FC<AccordionItemProps> = ({
   indicatorStyle,
   onSelected,
 }) => {
-  const theme = useXUITheme()
   const {
     variant,
     hideIndicator,
     disableAnimation,
     isCompact,
-    expandedKeys,
-    disabledKeys,
-    toggleItem,
-  } = useAccordionContext()
+    isExpanded,
+    isDisabled,
+    handlePress: togglePress,
+  } = useAccordionItemState(itemKey)
 
-  const resolvedItemKey = itemKey ?? ''
-  const isExpanded = resolvedItemKey ? expandedKeys.includes(resolvedItemKey) : false
-  const isDisabled = resolvedItemKey ? disabledKeys.includes(resolvedItemKey) : false
-
-  const [contentHeight, setContentHeight] = useState(0)
-  const [isMeasured, setIsMeasured] = useState(false)
-  const animatedHeight = useRef(new Animated.Value(isExpanded ? 1 : 0)).current
-  const animatedRotation = useRef(new Animated.Value(isExpanded ? 1 : 0)).current
-  const prevContentHeight = useRef(contentHeight)
-
-  useEffect(() => {
-    if (disableAnimation) {
-      animatedHeight.setValue(isExpanded ? 1 : 0)
-      animatedRotation.setValue(isExpanded ? 1 : 0)
-      return
-    }
-
-    Animated.parallel([
-      Animated.timing(animatedHeight, {
-        toValue: isExpanded ? 1 : 0,
-        duration: 300,
-        easing: Easing.bezier(0.4, 0.0, 0.2, 1),
-        useNativeDriver: false,
-      }),
-      Animated.timing(animatedRotation, {
-        toValue: isExpanded ? 1 : 0,
-        duration: 300,
-        easing: Easing.bezier(0.4, 0.0, 0.2, 1),
-        useNativeDriver: true,
-      }),
-    ]).start()
-  }, [isExpanded, disableAnimation, animatedHeight, animatedRotation])
-
-  useEffect(() => {
-    if (contentHeight > 0 && contentHeight !== prevContentHeight.current && isExpanded) {
-      prevContentHeight.current = contentHeight
-      if (!disableAnimation && isMeasured) {
-        Animated.timing(animatedHeight, {
-          toValue: 1,
-          duration: 200,
-          easing: Easing.out(Easing.ease),
-          useNativeDriver: false,
-        }).start()
-      }
-    }
-  }, [contentHeight, isExpanded, disableAnimation, isMeasured, animatedHeight])
-
-  const handlePress = useCallback(() => {
-    if (isDisabled || !resolvedItemKey) return
-    toggleItem(resolvedItemKey)
+  const handlePress = () => {
+    togglePress()
     onSelected?.(!isExpanded)
-  }, [isDisabled, toggleItem, resolvedItemKey, onSelected, isExpanded])
+  }
 
-  const onContentLayout = useCallback((event: LayoutChangeEvent) => {
-    const height = event.nativeEvent.layout.height
-    if (height > 0) {
-      setContentHeight(height)
-      setIsMeasured(true)
-    }
-  }, [])
+  const { onContentLayout, heightInterpolation, rotationInterpolation } =
+    useAccordionItemAnimation(isExpanded, disableAnimation)
 
-  const heightInterpolation = animatedHeight.interpolate({
-    inputRange: [0, 1],
-    outputRange: [0, contentHeight],
-  })
-
-  const rotationInterpolation = animatedRotation.interpolate({
-    inputRange: [0, 1],
-    outputRange: ['0deg', '90deg'],
-  })
-
-  const baseStyles = useMemo<ViewStyle>(() => {
-    const base: ViewStyle = {
-      overflow: 'hidden',
-    }
-
-    if (variant === 'splitted') {
-      base.paddingHorizontal = theme.spacing.md
-      base.backgroundColor = theme.colors.default.background
-      base.borderRadius = theme.borderRadius.md
-      base.marginBottom = theme.spacing.sm
-    } else if (variant === 'bordered') {
-      base.paddingHorizontal = theme.spacing.md
-    }
-
-    if (isDisabled) {
-      base.opacity = 0.4
-    }
-
-    return base
-  }, [variant, isDisabled, theme])
-
-  const triggerStyles = useMemo<ViewStyle>(() => {
-    const trigger: ViewStyle = {
-      flexDirection: 'row',
-      alignItems: 'center',
-      paddingVertical: isCompact ? theme.spacing.xs : theme.spacing.md,
-      gap: theme.spacing.md,
-    }
-
-    if (variant === 'light') {
-      trigger.paddingHorizontal = theme.spacing.sm
-    }
-
-    return trigger
-  }, [variant, isCompact, theme])
-
-  const titleTextStyle = useMemo<TextStyle>(() => {
-    return {
-      fontSize: isCompact ? theme.fontSizes.md : theme.fontSizes.lg,
-      fontWeight: theme.fontWeights.medium as TextStyle['fontWeight'],
-      color: theme.colors.foreground,
-    }
-  }, [isCompact, theme])
-
-  const subtitleTextStyle = useMemo<TextStyle>(() => {
-    return {
-      fontSize: theme.fontSizes.sm,
-      color: theme.colors.default.main,
-      marginTop: theme.spacing.xs,
-    }
-  }, [theme])
-
-  const contentContainerStyle = useMemo<ViewStyle>(() => {
-    return {
-      paddingBottom: isCompact ? theme.spacing.sm : theme.spacing.md,
-      paddingHorizontal: variant === 'light' ? theme.spacing.sm : 0,
-    }
-  }, [isCompact, variant, theme])
+  const {
+    baseStyles,
+    triggerStyles,
+    titleTextStyle,
+    subtitleTextStyle,
+    contentContainerStyle,
+    foregroundColor,
+  } = useAccordionItemStyles(variant, isCompact, isDisabled)
 
   return (
     <View style={[baseStyles, baseStyle]}>
@@ -193,15 +73,12 @@ export const AccordionItem: React.FC<AccordionItemProps> = ({
             ) : (
               title
             )}
-            {subtitle && (
-              <>
-                {typeof subtitle === 'string' ? (
-                  <Text style={[subtitleTextStyle, subtitleStyle]}>{subtitle}</Text>
-                ) : (
-                  subtitle
-                )}
-              </>
-            )}
+            {subtitle &&
+              (typeof subtitle === 'string' ? (
+                <Text style={[subtitleTextStyle, subtitleStyle]}>{subtitle}</Text>
+              ) : (
+                subtitle
+              ))}
           </View>
 
           {!hideIndicator && (
@@ -209,59 +86,33 @@ export const AccordionItem: React.FC<AccordionItemProps> = ({
               style={[
                 styles.indicator,
                 indicatorStyle,
-                {
-                  transform: [{ rotate: rotationInterpolation }],
-                },
+                { transform: [{ rotate: rotationInterpolation }] },
               ]}
             >
-              {indicator || <ChevronRightIcon color={theme.colors.foreground} />}
+              {indicator || <ChevronRightIcon color={foregroundColor} />}
             </Animated.View>
           )}
         </Pressable>
       </View>
 
-      <>
-        <View
-          style={{
-            position: 'absolute',
-            opacity: 0,
-            left: 0,
-            right: 0,
-            zIndex: -999,
-          }}
-          pointerEvents="none"
-        >
-          <View onLayout={onContentLayout} style={[contentContainerStyle, contentStyle]}>
-            {children}
-          </View>
+      <View style={styles.hiddenMeasure} pointerEvents="none">
+        <View onLayout={onContentLayout} style={[contentContainerStyle, contentStyle]}>
+          {children}
         </View>
-        <Animated.View
-          style={[
-            {
-              overflow: 'hidden',
-            },
-            disableAnimation
-              ? { height: isExpanded ? undefined : 0 }
-              : { height: heightInterpolation },
-          ]}
-        >
-          <View style={[contentContainerStyle, contentStyle]}>{children}</View>
-        </Animated.View>
-      </>
+      </View>
+
+      <Animated.View
+        style={[
+          styles.contentOverflow,
+          disableAnimation
+            ? { height: isExpanded ? undefined : 0 }
+            : { height: heightInterpolation },
+        ]}
+      >
+        <View style={[contentContainerStyle, contentStyle]}>{children}</View>
+      </Animated.View>
     </View>
   )
 }
 
 AccordionItem.displayName = 'AccordionItem'
-
-const styles = StyleSheet.create({
-  startContent: {
-    flexShrink: 0,
-  },
-  titleWrapper: {
-    flex: 1,
-  },
-  indicator: {
-    flexShrink: 0,
-  },
-})
