+++ packages/native/src/components/accordion/accordion.style.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  fullWidth: {
+    width: '100%',
+  },
+  divider: {
+    height: 1,
+    width: '100%',
+  },
+})
