+++ packages/native/src/components/accordion/accordion.tsx
@@ -1,30 +1,10 @@
-import React, { useState, useCallback, useMemo } from 'react'
-import { View, StyleSheet, type ViewStyle } from 'react-native'
-import { useXUITheme } from '../../core'
+import React, { useMemo } from 'react'
+import { View } from 'react-native'
 import { AccordionContext } from './accordion-context'
+import { styles } from './accordion.style'
+import { useAccordionStyles, useAccordionSelection } from './accordion.hook'
+import { getItemKey, isAccordionItem, normalizeElementKey } from './accordion.utils'
 import type { AccordionProps } from './accordion.type'
-import type { AccordionItemProps } from './accordion-item.type'
-import { AccordionItem } from './accordion-item'
-import { colors as palette } from '@xaui/core/palette'
-
-const getItemKey = (value: unknown, fallback: number) => {
-  if (value === null || value === undefined) return String(fallback)
-  if (typeof value === 'string' || typeof value === 'number') return String(value)
-  return String(fallback)
-}
-
-const normalizeElementKey = (value: unknown) => {
-  if (typeof value !== 'string') return value
-  return value.startsWith('.') ? value.slice(1) : value
-}
-
-const isAccordionItem = (
-  value: React.ReactNode
-): value is React.ReactElement<AccordionItemProps> =>
-  React.isValidElement(value) &&
-  (value.type === AccordionItem ||
-    (typeof value.type === 'function' &&
-      (value.type as { displayName?: string }).displayName === 'AccordionItem'))
 
 export const Accordion: React.FC<AccordionProps> = ({
   children,
@@ -42,53 +22,15 @@ export const Accordion: React.FC<AccordionProps> = ({
   itemStyle,
   onSelectionChange,
 }) => {
-  const theme = useXUITheme()
-  const [internalSelectedKeys, setInternalSelectedKeys] =
-    useState<string[]>(defaultSelectedKeys)
+  const { containerStyles, dividerColor, dividerOpacity } =
+    useAccordionStyles(variant, fullWidth)
 
-  const isControlled = selectedKeys !== undefined
-  const currentSelectedKeys = isControlled ? selectedKeys : internalSelectedKeys
-
-  const toggleItem = useCallback(
-    (key: string) => {
-      let newSelectedKeys: string[]
-
-      if (selectionMode === 'toggle') {
-        newSelectedKeys = currentSelectedKeys.includes(key) ? [] : [key]
-      } else {
-        newSelectedKeys = currentSelectedKeys.includes(key)
-          ? currentSelectedKeys.filter(currentKey => currentKey !== key)
-          : [...currentSelectedKeys, key]
-      }
-
-      if (!isControlled) {
-        setInternalSelectedKeys(newSelectedKeys)
-      }
-
-      onSelectionChange?.(newSelectedKeys)
-    },
-    [selectionMode, currentSelectedKeys, isControlled, onSelectionChange]
+  const { currentSelectedKeys, toggleItem } = useAccordionSelection(
+    selectionMode,
+    selectedKeys,
+    defaultSelectedKeys,
+    onSelectionChange
   )
-
-  const containerStyles = useMemo<ViewStyle>(() => {
-    const styles: ViewStyle = {}
-
-    if (fullWidth) {
-      styles.width = '100%'
-    }
-
-    if (variant === 'bordered') {
-      styles.borderWidth = theme.borderWidth.md
-      styles.borderColor = palette.gray[200]
-      styles.borderRadius = theme.borderRadius.md
-      styles.marginHorizontal = theme.spacing.sm
-      styles.overflow = 'hidden'
-    } else if (variant === 'light') {
-      styles.paddingHorizontal = theme.spacing.sm
-    }
-
-    return styles
-  }, [variant, fullWidth, theme])
 
   const contextValue = useMemo(
     () => ({
@@ -122,14 +64,9 @@ export const Accordion: React.FC<AccordionProps> = ({
           const isLast = index === childrenArray.length - 1
           const showBottomDivider =
             (showDivider || variant === 'bordered') && !isLast && variant !== 'splitted'
-          const dividerColor =
-            variant === 'bordered' ? palette.gray[200] : theme.colors.default.foreground
-          const dividerOpacity = variant === 'bordered' ? 1 : 0.2
+
           const resolvedChildKey = isAccordionItem(child)
-            ? getItemKey(
-                child.props.itemKey ?? normalizeElementKey(child.key),
-                index
-              )
+            ? getItemKey(child.props.itemKey ?? normalizeElementKey(child.key), index)
             : getItemKey(
                 React.isValidElement(child) ? normalizeElementKey(child.key) : undefined,
                 index
@@ -158,10 +95,3 @@ export const Accordion: React.FC<AccordionProps> = ({
     </AccordionContext.Provider>
   )
 }
-
-const styles = StyleSheet.create({
-  divider: {
-    height: 1,
-    width: '100%',
-  },
-})
